+++ lib/src/onboarding/sbb_onboarding_builder_delegate.dart
@@ -2,6 +2,9 @@ import 'package:flutter/material.dart';
 
 import 'sbb_onboarding_card.dart';
 
+@Deprecated(
+  'SBBOnboarding is deprecated since it is no component in the official Design System. Will be removed in v5.',
+)
 abstract class SBBOnboardingBuilderDelegate {
   Widget buildStartPage(BuildContext context, VoidCallback onStartOnboarding, VoidCallback onFinish);