+++ test/sbb_tab_bar_test.dart
@@ -18,22 +18,27 @@ void main() {
     final items4 = Iterable.generate(4, (i) => _DemoItem('${i}_d', icons[i % icons.length])).toList();
     final items5 = Iterable.generate(5, (i) => _DemoItem('${i}_e', icons[i % icons.length])).toList();
     final items6 = Iterable.generate(6, (i) => _DemoItem('${i}_f', icons[i % icons.length])).toList();
+    final controller1 = SBBTabBarController(items3, items3[0])
+      ..setBadges([
+        SBBTabBarItemBadge(
+          id: '1_c',
+          badge: SBBTabBarBadgeIcon(badgeIcon: SBBBadgeIconData.checkmark),
+        ),
+      ]);
+    final controller2 = SBBTabBarController(items4, items4.last)
+      ..setBadges([
+        SBBTabBarItemBadge(
+          id: '2_d',
+          badge: SBBTabBarBadgeText(labelText: '99+'),
+        ),
+      ]);
+
     final widget = Column(
       children: [
         SBBTabBar.items(items: items, onTabChanged: (tab) async {}, onTap: (tab) {}),
         SBBTabBar.items(items: items2, onTabChanged: (tab) async {}, initialItem: items2.last, onTap: (tab) {}),
-        SBBTabBar.items(
-          items: items3,
-          onTabChanged: (tab) async {},
-          initialItem: items3.skip(1).first,
-          onTap: (tab) {},
-        ),
-        SBBTabBar.items(
-          items: items4,
-          onTabChanged: (tab) async {},
-          initialItem: items4.skip(2).first,
-          onTap: (tab) {},
-        ),
+        SBBTabBar.controller(controller: controller1, onTabChanged: (tab) async {}, onTap: (tab) {}),
+        SBBTabBar.controller(controller: controller2, onTabChanged: (tab) async {}, onTap: (tab) {}),
         SBBTabBar.items(
           items: items5,
           onTabChanged: (tab) async {},